1. nanf("")를 NAN 매크로로 통일 2. 기타 백업

+++ Includes/Indicators/StandardDeviation.hpp
@@ -16,11 +16,11 @@ class StandardDeviation final : public Indicator {
   size_t sizet_period_;
 
   size_t count_;
-  long double sum_;
-  long double sum_sq_;
+  double sum_;
+  double sum_sq_;
   bool can_calc_;
 
-  vector<long double> buffer_;
+  vector<double> buffer_;
   size_t buffer_idx_;
 
   void Initialize() override;
